fix(Typography): expose text styles
