+++ packages/blade/src/components/Typography/BaseText/BaseText.d.ts
@@ -1,0 +1,2 @@
+export { BaseText } from './BaseText.web';
+export * from './BaseTextTypes';
