+++ packages/blade/src/components/Typography/BaseText/BaseTextTypes.ts
@@ -1,0 +1,48 @@
+import type { Theme } from '~components/BladeProvider';
+import type { DotNotationColorStringToken } from '~src/_helpers/types';
+import type { Feedback } from '~tokens/theme/theme';
+import type { AccessibilityProps } from '~utils';
+
+type FeedbackColors = `feedback.text.${DotNotationColorStringToken<
+  Theme['colors']['feedback']['text']
+>}`;
+type FeedbackActionColors = `feedback.${Feedback}.action.text.${DotNotationColorStringToken<
+  Theme['colors']['feedback'][Feedback]['action']['text']
+>}`;
+type SurfaceColors = `surface.text.${DotNotationColorStringToken<
+  Theme['colors']['surface']['text']
+>}`;
+type ActionColors = `action.text.${DotNotationColorStringToken<Theme['colors']['action']['text']>}`;
+
+export type BaseTextProps = {
+  id?: string;
+  color?: ActionColors | FeedbackColors | SurfaceColors | FeedbackActionColors;
+  fontFamily?: keyof Theme['typography']['fonts']['family'];
+  fontSize?: keyof Theme['typography']['fonts']['size'];
+  fontWeight?: keyof Theme['typography']['fonts']['weight'];
+  fontStyle?: 'italic' | 'normal';
+  textDecorationLine?: 'line-through' | 'none' | 'underline';
+  lineHeight?: keyof Theme['typography']['lineHeights'];
+  as?: 'code' | 'h1' | 'h2' | 'h3' | 'h4' | 'h5' | 'h6' | 'p' | 'span';
+  textAlign?: 'center' | 'justify' | 'left' | 'right';
+  truncateAfterLines?: number;
+  className?: string;
+  style?: React.CSSProperties;
+  children: React.ReactNode;
+  accessibilityProps?: Partial<AccessibilityProps>;
+  numberOfLines?: number;
+};
+
+export type StyledBaseTextProps = Pick<
+  BaseTextProps,
+  | 'color'
+  | 'fontFamily'
+  | 'fontSize'
+  | 'fontWeight'
+  | 'fontStyle'
+  | 'textDecorationLine'
+  | 'lineHeight'
+  | 'as'
+  | 'textAlign'
+  | 'numberOfLines'
+> & { theme: Theme };
