--- packages/blade/src/components/Typography/BaseText/StyledBaseText.d.ts
@@ -1,17 +1,0 @@
-import type { AriaRoles } from '~utils';
-
-export { default } from './StyledBaseText.web';
-
-export type StyledBaseTextProps = {
-  color: string;
-  fontFamily: string;
-  fontSize: string;
-  fontWeight: number;
-  fontStyle?: 'italic' | 'normal';
-  textDecorationLine?: 'line-through' | 'none' | 'underline';
-  lineHeight: string;
-  as?: 'code' | 'h1' | 'h2' | 'h3' | 'h4' | 'h5' | 'h6' | 'p' | 'span';
-  textAlign?: 'center' | 'justify' | 'left' | 'right';
-  numberOfLines?: number;
-  accessibilityRole?: AriaRoles;
-};
