--- packages/blade/src/components/Typography/BaseText/StyledBaseText.web.tsx
@@ -1,28 +1,0 @@
-import styled from 'styled-components';
-import getBaseTextStyles from './getBaseTextStyles';
-import type { StyledBaseTextProps } from './StyledBaseText.d';
-
-const StyledBaseText = styled.div<StyledBaseTextProps>(
-  ({
-    color,
-    fontFamily,
-    fontSize,
-    fontWeight,
-    fontStyle,
-    textDecorationLine,
-    lineHeight,
-    textAlign,
-  }) =>
-    getBaseTextStyles({
-      color,
-      fontFamily,
-      fontSize,
-      fontWeight,
-      fontStyle,
-      textDecorationLine,
-      lineHeight,
-      textAlign,
-    }),
-);
-
-export default StyledBaseText;
