+++ packages/blade/src/components/Typography/BaseText/getBaseTextStyles.ts
@@ -1,25 +1,36 @@
-import type { StyledBaseTextProps } from './StyledBaseText.d';
+import type { CSSObject } from 'styled-components';
+import type { StyledBaseTextProps } from './BaseTextTypes';
+import { getIn, makeTypographySize } from '~utils';
 
 const getBaseTextStyles = ({
-  color,
-  fontFamily,
-  fontSize,
-  fontWeight,
+  color = 'surface.text.normal.lowContrast',
+  fontFamily = 'text',
+  fontSize = 200,
+  fontWeight = 'regular',
   fontStyle = 'normal',
   textDecorationLine = 'none',
-  lineHeight,
-  textAlign,
-}: StyledBaseTextProps): string => `
-  color: ${color};
-  font-family: ${fontFamily};
-  font-size: ${fontSize};
-  font-weight: ${fontWeight};
-  font-style: ${fontStyle};
-  text-decoration-line: ${textDecorationLine};
-  line-height: ${lineHeight};
-  ${textAlign ? `text-align: ${textAlign};` : ''};
-  margin: 0;
-  padding: 0;
-`;
+  lineHeight = 'l',
+  textAlign = 'left',
+  theme,
+}: StyledBaseTextProps): CSSObject => {
+  const textColor = getIn(theme.colors, color);
+  const themeFontFamily = theme.typography.fonts.family[fontFamily];
+  const themeFontSize = makeTypographySize(theme.typography.fonts.size[fontSize]);
+  const themeFontWeight = theme.typography.fonts.weight[fontWeight];
+  const themeLineHeight = makeTypographySize(theme.typography.lineHeights[lineHeight]);
+
+  return {
+    color: textColor,
+    fontFamily: themeFontFamily,
+    fontSize: themeFontSize,
+    fontWeight: themeFontWeight,
+    fontStyle,
+    textDecorationLine,
+    lineHeight: themeLineHeight,
+    textAlign,
+    margin: 0,
+    padding: 0,
+  };
+};
 
 export default getBaseTextStyles;
