+++ packages/blade/src/components/Typography/BaseText/index.ts
@@ -1,2 +1,1 @@
-export { default } from './BaseText';
 export * from './BaseText';
