+++ packages/blade/src/components/Typography/Heading/Heading.tsx
@@ -1,6 +1,6 @@
 /* eslint-disable @typescript-eslint/no-unnecessary-type-assertion */
 import type { ReactElement } from 'react';
-import BaseText from '../BaseText';
+import { BaseText } from '../BaseText';
 import type { BaseTextProps } from '../BaseText';
 import type { ColorContrast, ColorContrastTypes, TextTypes } from '~tokens/theme/theme';
 import { getPlatformType } from '~utils';
@@ -83,7 +83,7 @@ const getProps = <T extends { variant: HeadingVariant }>({
   return props;
 };
 
-const Heading = <T extends { variant: HeadingVariant }>({
+export const Heading = <T extends { variant: HeadingVariant }>({
   variant = 'small',
   type = 'normal',
   weight = 'bold',
@@ -93,5 +93,3 @@ const Heading = <T extends { variant: HeadingVariant }>({
   const props = getProps({ variant, type, weight, contrast });
   return <BaseText {...props}>{children}</BaseText>;
 };
-
-export default Heading;
