+++ packages/blade/src/components/Typography/Heading/__tests__/Heading.native.test.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Heading from '../';
+import { Heading } from '../';
 import renderWithTheme from '~src/_helpers/testing/renderWithTheme.native';
 
 beforeAll(() => jest.spyOn(console, 'error').mockImplementation());
