+++ packages/blade/src/components/Typography/Heading/index.ts
@@ -1,2 +1,1 @@
-export { default } from './Heading';
 export * from './Heading';
