+++ packages/blade/src/components/Typography/Text/index.ts
@@ -1,2 +1,1 @@
-export { default } from './Text';
 export * from './Text';
