+++ packages/blade/src/components/Typography/Title/__tests__/Title.native.test.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Title from '../';
+import { Title } from '../';
 import renderWithTheme from '~src/_helpers/testing/renderWithTheme.native';
 
 beforeAll(() => jest.spyOn(console, 'error').mockImplementation());
