+++ packages/blade/src/components/Typography/Title/index.ts
@@ -1,2 +1,1 @@
-export { default } from './Title';
 export * from './Title';
