+++ packages/blade/src/components/Typography/index.ts
@@ -1,6 +1,6 @@
-export { default as Title } from './Title';
+export { Title } from './Title';
 export * from './Title';
-export { default as Heading } from './Heading';
+export { Heading } from './Heading';
 export * from './Heading';
-export { default as Text } from './Text';
+export { Text } from './Text';
 export * from './Text';
